refactor: split analysis/subagents/file_queue.py into focused modules
Extract state persistence (atomic JSON read/write, file locking) into
_queue_state.py (75 lines). Slim file_queue.py to 98 lines with
re-exports preserving all existing import paths.
